requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
sqlparse==0.5.0
